+++ selene/support/__init__.py
@@ -1,3 +1,3 @@
 # TODO: consider renaming support to _support to emphasize its experimental nature
 
-from . import _logging, shared
+from . import _logging